Add laba4 front

+++ Sem 3/Веб/laba4/laba4_back/src/main/java/org/michael/laba4_back/controller/PointController.java
@@ -33,24 +33,35 @@ public class PointController {
             HttpServletRequest request
     ) {
         AuthToken token = authTokenRepository.findByToken(request.getHeader("auth-token"));
+        if (token == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(getMessage("Wrong auth token"));
+        if (token.isExpired()) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(getMessage("Auth token expired"));
         if (!Point.isValid(x, y, r)) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(getMessage("Invalid values"));
         }
         Point point = new Point(x, y, r, Point.isHit(x, y, r), LocalDateTime.now(), token.getUser());
         pointRepository.save(point);
-        return ResponseEntity.ok(gson.toJson(pointRepository.findByUserId(token.getUser().getId())));
+        return ResponseEntity.ok(gson.toJson(point));
     }
 
     @GetMapping(value = "/point", produces = "application/json")
     public ResponseEntity<String> getPoints(HttpServletRequest request) {
         AuthToken token = authTokenRepository.findByToken(request.getHeader("auth-token"));
+        if (token == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(getMessage("Wrong auth token"));
+        if (token.isExpired()) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(getMessage("Auth token expired"));
         return ResponseEntity.ok(gson.toJson(pointRepository.findByUserId(token.getUser().getId())));
     }
 
     @DeleteMapping(value = "/point", produces = "application/json")
     public ResponseEntity<String> deletePoints(HttpServletRequest request) {
         AuthToken token = authTokenRepository.findByToken(request.getHeader("auth-token"));
+        if (token == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(getMessage("Wrong auth token"));
+        if (token.isExpired()) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(getMessage("Auth token expired"));
         pointRepository.deleteByUserId(token.getUser().getId());
         return ResponseEntity.ok(getMessage("Points deleted"));
+    }
+
+    @RequestMapping(value = "/point", method = RequestMethod.OPTIONS)
+    public ResponseEntity<?> options() {
+        return ResponseEntity.ok().build();
     }
 }
